requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
spacy==3.8.14
beautifulsoup4==4.14.3
requests==2.32.5
extruct==0.18.0
w3lib==2.4.1
python-multipart==0.0.26
pytesseract==0.3.13
pillow==12.2.0
pdfminer.six==20251230
lxml==6.0.2